Implemented song based switching dimensions
Joreys 2 songs were implemented. additionally the two songs play in the correct dimension only. Halfways through the song (or however many seconds through) The background Dimensions will shift. and the player will no longer be able to swap to the previous two dimensions. There are a few bugs however: The current bg switching must get through its original two sets of dimensions before moving on to the new sets. While its in these two sets, the player cannot switch to the other dimension because they are stuck in the new dimension. I will have a fix for this shortly.